% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_io.R
\name{Bed2Granges}
\alias{Bed2Granges}
\title{Import BED as Genomic Ranges}
\usage{
Bed2Granges(bedFile)
}
\arguments{
\item{bedFile}{Path to BED file.}
}
\value{
A \code{GRanges} object.
}
\description{
Import BED as Genomic Ranges
}
