% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_legacy_migrated.R
\name{CalcStatByGrp}
\alias{CalcStatByGrp}
\title{Comprehensive Imprinting Analysis for imprintomeR}
\usage{
CalcStatByGrp(dat, meta, low_cutoff = 0.3, high_cutoff = 0.7)
}
\arguments{
\item{betaFile}{Character. file of beta values (Probes as rownames, Samples as colnames).}

\item{metaFile}{Character. file of sample information.}

\item{probeset}{Character. the name of a predefined imprinting probeset.}

\item{ids_cutoff}{Numeric. The IDS distance threshold to call an EpiMutation (Default = 0.2).}

\item{region_name}{Character. Optional label for the genomic locus (e.g., "11p15").}
}
\value{
A data frame with metrics, directional interpretations, and mutation calls.
}
\description{
Comprehensive Imprinting Analysis for imprintomeR
}
