% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_legacy_migrated.R
\name{Calculate_Stat_by_range_v6}
\alias{Calculate_Stat_by_range_v6}
\title{Calculate ImprintAge and Age Acceleration}
\usage{
Calculate_Stat_by_range_v6(value, low_cutoff = 0.3, high_cutoff = 0.7)
}
\arguments{
\item{beta_matrix}{Matrix of beta values (rows = CpGs, cols = samples)}

\item{icr_sites}{Vector of CpG IDs (cgXXXX) that reside within ICRs}

\item{global_clocks_age}{Numeric vector of previously calculated BioAge (e.g., Horvath Age)}

\item{trained_weights}{A named vector of weights for ICR sites (from a training set)}
}
\value{
A dataframe with ImprintAge and Acceleration scores
}
\description{
Calculate ImprintAge and Age Acceleration
}
