% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_io.R
\name{CheckHeader}
\alias{CheckHeader}
\title{Detect Presence of Header in Probe List File}
\usage{
CheckHeader(file_path)
}
\arguments{
\item{file_path}{Input text file path.}
}
\value{
Logical; \code{TRUE} when a header line is detected.
}
\description{
Detect Presence of Header in Probe List File
}
