% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_io.R
\name{CheckProbeIDs}
\alias{CheckProbeIDs}
\title{Check Probe ID Format}
\usage{
CheckProbeIDs(ids)
}
\arguments{
\item{ids}{Character vector of probe IDs.}
}
\value{
Logical vector; \code{TRUE} for IDs matching \code{cg########}.
}
\description{
Check Probe ID Format
}
