% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_io.R
\name{CheckSentrixID}
\alias{CheckSentrixID}
\title{Validate Illumina Sentrix IDs}
\usage{
CheckSentrixID(ids)
}
\arguments{
\item{ids}{Character vector of Sentrix IDs.}
}
\value{
Logical vector indicating valid IDs.
}
\description{
Validate Illumina Sentrix IDs
}
