% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethQcSet-methods.R
\name{computeQC,MethQcSet-method}
\alias{computeQC,MethQcSet-method}
\title{Compute QC Metrics on a MethQcSet}
\usage{
\S4method{computeQC}{MethQcSet}(x, detection_pval = NULL, pcutoff = 0.05, ...)
}
\arguments{
\item{x}{A \code{MethQcSet} object.}

\item{detection_pval}{Optional matrix of detection p-values. If NULL, uses x@detection_pval.}

\item{pcutoff}{Numeric, maximum mean detection p-value cutoff (default: 0.05).}


\item{...}{Additional arguments (reserved for future use).}
}
\value{
A \code{MethQcSet} object with updated qc_tables and qc_params.
}
\description{
Internal method. Compute QC metrics from beta values and detection p-values,
storing results in qc_tables. Called internally by \code{runMethQC()}.
}
\keyword{internal}
