% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethQcSet-merge.R
\name{find_intersection}
\alias{find_intersection}
\title{Find Common Samples or Probes Between MethQcSet Objects}
\usage{
find_intersection(x, y, by = c("samples", "probes"))
}
\arguments{
\item{x}{A \code{MethQcSet} object.}

\item{y}{A \code{MethQcSet} object.}

\item{by}{Character, "samples" (default) to find common samples, or "probes" for common probes.}
}
\value{
Character vector of common identifiers (sample names or probe IDs).
}
\description{
Identify common samples or probes between two \code{MethQcSet} objects.
}
