% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_legacy_migrated.R
\name{Imprintome_Classifier}
\alias{Imprintome_Classifier}
\title{Calculate Epigenetic Stochasticity (ImpVar)}
\usage{
Imprintome_Classifier(
  betaFile,
  metaFile,
  probeset = "selected",
  prefix = NULL,
  low_cutoff = 0.3,
  high_cutoff = 0.7
)
}
\arguments{
\item{beta_matrix}{A matrix of beta values (rows = CpGs, cols = samples)}

\item{cpg_annos}{A data frame or GRanges with CpG positions (must match beta_matrix rows)}

\item{icr_regions}{A GRanges object defining the start/end of ICRs}
}
\value{
A data frame of Variance scores per ICR per sample
}
\description{
Calculate Epigenetic Stochasticity (ImpVar)
}
