% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImprintomeSet-accessors.R
\name{ImprintomeSet-accessors}
\alias{ImprintomeSet-accessors}
\alias{beta,ImprintomeSet-method}
\alias{beta<-,ImprintomeSet-method}
\alias{meta,ImprintomeSet-method}
\alias{meta<-,ImprintomeSet-method}
\alias{probeset,ImprintomeSet-method}
\alias{probeset<-,ImprintomeSet-method}
\alias{genome,ImprintomeSet-method}
\alias{genome<-,ImprintomeSet-method}
\alias{assay,ImprintomeSet-method}
\alias{assay<-,ImprintomeSet-method}
\alias{results,ImprintomeSet-method}
\alias{results<-,ImprintomeSet-method}
\alias{plots,ImprintomeSet-method}
\alias{plots<-,ImprintomeSet-method}
\title{Accessors for ImprintomeSet}
\usage{
\S4method{beta}{ImprintomeSet}(x)

\S4method{beta}{ImprintomeSet}(x) <- value

\S4method{meta}{ImprintomeSet}(x)

\S4method{meta}{ImprintomeSet}(x) <- value

\S4method{probeset}{ImprintomeSet}(x)

\S4method{probeset}{ImprintomeSet}(x) <- value

\S4method{genome}{ImprintomeSet}(x)

\S4method{genome}{ImprintomeSet}(x) <- value

\S4method{assay}{ImprintomeSet}(x)

\S4method{assay}{ImprintomeSet}(x) <- value

\S4method{results}{ImprintomeSet}(x)

\S4method{results}{ImprintomeSet}(x) <- value

\S4method{plots}{ImprintomeSet}(x)

\S4method{plots}{ImprintomeSet}(x) <- value
}
\arguments{
\item{x}{An \code{ImprintomeSet} object.}

\item{value}{Replacement value for a slot.}
}
\value{
Getter methods return the requested slot value.
Replacement methods return an updated \code{ImprintomeSet} object.
}
\description{
Bioconductor-style accessors for reading and updating \code{ImprintomeSet} slots.
}
