% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethQcSet-accessors.R
\name{MethQcSet-accessors}
\alias{MethQcSet-accessors}
\alias{meta,MethQcSet-method}
\alias{meta<-,MethQcSet-method}
\alias{platform,MethQcSet-method}
\alias{platform<-,MethQcSet-method}
\alias{beta,MethQcSet-method}
\alias{beta<-,MethQcSet-method}
\alias{detection_pval,MethQcSet-method}
\alias{detection_pval<-,MethQcSet-method}
\alias{qc_tables,MethQcSet-method}
\alias{qc_tables<-,MethQcSet-method}
\alias{aggregation_status,MethQcSet-method}
\alias{aggregation_status<-,MethQcSet-method}
\alias{qc_params,MethQcSet-method}
\alias{qc_params<-,MethQcSet-method}
\alias{statistics,MethQcSet-method}
\alias{statistics<-,MethQcSet-method}
\title{Accessors for MethQcSet}
\usage{
\S4method{meta}{MethQcSet}(x)

\S4method{meta}{MethQcSet}(x) <- value

\S4method{platform}{MethQcSet}(x)

\S4method{platform}{MethQcSet}(x) <- value

\S4method{beta}{MethQcSet}(x)

\S4method{beta}{MethQcSet}(x) <- value

\S4method{detection_pval}{MethQcSet}(x)

\S4method{detection_pval}{MethQcSet}(x) <- value

\S4method{qc_tables}{MethQcSet}(x)

\S4method{qc_tables}{MethQcSet}(x) <- value

\S4method{aggregation_status}{MethQcSet}(x)

\S4method{aggregation_status}{MethQcSet}(x) <- value

\S4method{qc_params}{MethQcSet}(x)

\S4method{qc_params}{MethQcSet}(x) <- value

\S4method{statistics}{MethQcSet}(x)

\S4method{statistics}{MethQcSet}(x) <- value
}
\arguments{
\item{x}{A \code{MethQcSet} object.}

\item{value}{Replacement value for a slot.}
}
\value{
Getter methods return the requested slot value.
Replacement methods return an updated \code{MethQcSet} object.
}
\description{
Bioconductor-style accessors for reading and updating \code{MethQcSet} slots.
}
