% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethQcSet-io.R
\name{MethQcSet-io}
\alias{MethQcSet-io}
\title{Platform Detection and QC Factory for MethQcSet}
\description{
Functions for detecting array platform from IDAT files and creating
QC-processed \code{MethQcSet} objects from raw IDAT data.
}
