% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethQcSet-merge.R
\name{MethQcSet-merge}
\alias{MethQcSet-merge}
\title{Merge and Intersection Methods for MethQcSet}
\description{
Methods for combining multiple MethQcSet objects with synchronized metadata and beta alignment.
}
