% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethQcSet-methods.R
\name{MethQcSet-methods}
\alias{MethQcSet-methods}
\title{QC Computation and Aggregation Methods for MethQcSet}
\description{
Methods for computing QC metrics and handling platform-specific probe aggregation.
}
