% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_io.R
\name{ParseBedFile}
\alias{ParseBedFile}
\title{Parse, Sort, and Normalize BED File}
\usage{
ParseBedFile(bedFile)
}
\arguments{
\item{bedFile}{Path to BED file.}
}
\value{
Character vector of BED regions.
}
\description{
Ensures \code{chr} prefix, sorts by genomic coordinates, writes a sorted BED copy,
and returns BED regions as \code{chr:start-end} strings.
}
