% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_plotting.R
\name{PlotRainfall}
\alias{PlotRainfall}
\title{Comprehensive Imprinting Analysis for imprintomeR}
\usage{
PlotRainfall(
  beta,
  sampleID,
  title = "Imprinting Rainfall Plot",
  probeset = c("selected", "NanoImprint", "Joshi", "Court", "Rosenski", "Jima", "chr11p15"),
  outFile = NULL
)
}
\arguments{
\item{probeset}{Character. the name of a predefined imprinting probeset.}

\item{betaFile}{Character. file of beta values (Probes as rownames, Samples as colnames).}

\item{metaFile}{Character. file of sample information.}

\item{ids_cutoff}{Numeric. The IDS distance threshold to call an EpiMutation (Default = 0.2).}
}
\value{
A data frame with metrics, directional interpretations, and mutation calls.
}
\description{
Comprehensive Imprinting Analysis for imprintomeR
}
