% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_scoring.R
\name{run_pic_signature_finder}
\alias{run_pic_signature_finder}
\title{Pan-Imprint Clustering (PIC) Module}
\usage{
run_pic_signature_finder(icr_beta_matrix, results_dir = "PIC_Output")
}
\arguments{
\item{icr_beta_matrix}{Matrix of beta values filtered for ICR sites only}

\item{results_dir}{Directory to save consensus plots}
}
\value{
A list containing cluster assignments and the signature heatmap
}
\description{
Pan-Imprint Clustering (PIC) Module
}
