% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImprintomeSet-class.R
\name{ImprintomeSet-class}
\alias{ImprintomeSet-class}
\title{ImprintomeSet S4 Class}
\description{
Formal container for imprintomeR inputs and derived outputs.
}
\section{Slots}{

\describe{
\item{\code{beta}}{Matrix or data.frame of beta values (rows = probes, columns = samples).}

\item{\code{meta}}{data.frame of sample metadata. Must include \code{SAMPLE_NAME}.}

\item{\code{probeset}}{data.frame or list containing probeset annotation.}

\item{\code{genome}}{Character scalar indicating genome build.}

\item{\code{assay}}{Character scalar indicating array assay.}

\item{\code{results}}{Named list of derived result tables/objects.}

\item{\code{plots}}{Named list of generated plots.}
}}

